requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
execnet==2.0.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
PyYAML==6.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.2.1
